pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sonarqube-scanner"
version = "0.2.0"
description = "Scan Cosmo-Tech platform repositories with SonarQube"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Diamantis Sellis", email = "diamantis.sellis@cosmotech.com"}
]
dependencies = [
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "click>=8.1.8",
    "ruff>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
sonar-scan = "sonarqube_scanner.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/sonarqube_scanner"]
